requirements.txt
Django==3.2.6
psycopg2-binary==2.9.1
djangorestframework==3.12.4
django-money==2.0.1
django-filter==2.4.0
pytest==6.2.4
pytest-django==4.4.0
django-debug-toolbar==3.2.2
flake8==3.9.2
flake8-coding==1.3.2
flake8-quotes==3.3.0
flake8-isort==4.0.0
celery==5.1.2
redis==3.4.1
django-environ==0.4.5
django-crum==0.7.9
Pillow==8.3.1
requests==2.26.0
pytest-cov==2.12.1
pytest-xdist==2.3.0
coverage==5.5
django-rest-swagger==2.2.0